Fix one-hot basin encoding (#150)
The one-hot basin encoding used a single tensor (i.e., a single location
in memory) for all encodings. Hence, when a minibatch was created and
__getitem__ was called several times, the one-hot encoding of the
previous __getitem__ call would be overwritten by the encoding of the
current call.

Now, we create a new tensor in each __getitem__ to avoid this problem.